Move ErrorPage to layouts (not a component)?
